Changed all press values to floats. No need for such high precision.

+++ Assets/MixedRealityToolkit/InputSystem/MixedRealityInputManager.cs
@@ -774,7 +774,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
         }
 
         /// <inheritdoc />
-        public void RaiseOnInputPressed(IMixedRealityInputSource source, double pressAmount)
+        public void RaiseOnInputPressed(IMixedRealityInputSource source, float pressAmount)
         {
             // Create input event
             inputPressedEventData.Initialize(source, pressAmount);
@@ -784,7 +784,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
         }
 
         /// <inheritdoc />
-        public void RaiseOnInputPressed(IMixedRealityInputSource source, KeyCode keyCode, double pressAmount)
+        public void RaiseOnInputPressed(IMixedRealityInputSource source, KeyCode keyCode, float pressAmount)
         {
             // Create input event
             inputPressedEventData.Initialize(source, keyCode, pressAmount);
@@ -794,7 +794,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
         }
 
         /// <inheritdoc />
-        public void RaiseOnInputPressed(IMixedRealityInputSource source, InputType inputType, double pressAmount)
+        public void RaiseOnInputPressed(IMixedRealityInputSource source, InputType inputType, float pressAmount)
         {
             // Create input event
             inputPressedEventData.Initialize(source, inputType, pressAmount);
@@ -804,7 +804,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
         }
 
         /// <inheritdoc />
-        public void RaiseOnInputPressed(IMixedRealityInputSource source, Handedness handedness, double pressAmount)
+        public void RaiseOnInputPressed(IMixedRealityInputSource source, Handedness handedness, float pressAmount)
         {
             // Create input event
             inputPressedEventData.Initialize(source, handedness, pressAmount);
@@ -814,7 +814,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
         }
 
         /// <inheritdoc />
-        public void RaiseOnInputPressed(IMixedRealityInputSource source, Handedness handedness, KeyCode keyCode, double pressAmount)
+        public void RaiseOnInputPressed(IMixedRealityInputSource source, Handedness handedness, KeyCode keyCode, float pressAmount)
         {
             // Create input event
             inputPressedEventData.Initialize(source, handedness, keyCode, pressAmount);
@@ -824,7 +824,7 @@ namespace Microsoft.MixedReality.Toolkit.InputSystem
         }
 
         /// <inheritdoc />
-        public void RaiseOnInputPressed(IMixedRealityInputSource source, Handedness handedness, InputType inputType, double pressAmount)
+        public void RaiseOnInputPressed(IMixedRealityInputSource source, Handedness handedness, InputType inputType, float pressAmount)
         {
             // Create input event
             inputPressedEventData.Initialize(source, handedness, inputType, pressAmount);
